+++ app/desktop/MyPal.Desktop/ViewModels/ViewModelBase.cs
@@ -1,7 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace MyPal.Desktop.ViewModels;
-
-public class ViewModelBase : ObservableObject
-{
-}